requirements.txt
boto3==1.34.0
fastapi==0.115.12
Flask==3.1.1
flask_socketio==5.5.1
langchain_community==0.3.24
langchain_core==0.3.61
neo4j==5.28.1
pydantic==2.11.5
python-dotenv==1.1.0
uvicorn==0.34.2
